% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjolnir1_RAN.R
\name{mjolnir1_RAN}
\alias{mjolnir1_RAN}
\title{RAN: Demultiplex and trim primers from raw FASTQ files}
\usage{
mjolnir1_RAN(
  R1_filenames = "",
  lib_prefix = "",
  experiment = NULL,
  primer_F = "GGWACWRGWTGRACWNTNTAYCCYCC",
  primer_R = "TANACYTCNGGRTGNCCRAARAAYCA",
  cores = 1,
  R1_motif = "_R1",
  R2_motif = "_R2",
  metadata = "",
  multilane = FALSE,
  tag_error = 0,
  primer_error = 0.1,
  original_samples = "original_samples",
  mjolnir_agnomens = "mjolnir_agnomens",
  commands_file = "commands_runned_RAN.txt",
  only_commands = FALSE,
  ...
)
}
\arguments{
\item{R1_filenames}{Character vector with the names of the forward fastq or
fastq.gz files. Only needed for multiplexed libraries.}

\item{lib_prefix}{Character vector. Acronym for each sequencing library. This
acronym must be of 4 characters in capital letters. Do not mix up library and
experiment acronyms. The latter will be required in following steps. However
they can be the same.}

\item{experiment}{Character string. Acronym for the experiment. This
acronym must be of 4 characters in capital letters. Do not mix up library and
experiment acronyms. However they can be the same.}

\item{primer_F}{Character string of the Forward primer. Necessary when
samples are already demultiplexed.}

\item{primer_R}{Character string of the Reverse primer. Necessary when
samples are already demultiplexed.}

\item{cores}{Numeric. Number threads to run cutadapt.}

\item{R1_motif}{Character string that distinguishes
the forward line file from
the reverse.}

\item{R2_motif}{Character string that distinguishes
the reverse line file from
the forward.}

\item{multilane}{Logical. If FALSE, the function will
consider that the each library
has only one sequencing lane. If TRUE,
only one library is processed at a time.}

\item{tag_error}{Numeric. From 0 (no errors allowed) to 1. will determine the
proportion of nucleotides that can be different in the sample tags.}

\item{primer_error}{Numeric. From 0 (no errors allowed) to 1.
will determine the
proportion of nucleotides that can be different in the primers.}

\item{original_samples}{Character string. Name of the column in the metadata
table that contains the original sample names.}

\item{mjolnir_agnomens}{Character string. Name of the column in the metadata
table that contains the mjolnir agnomen names.}

\item{commands_file}{Character string. Name of the file where all commands
will be written. If NULL or missing, commands will not be recorded.}

\item{only_commands}{Logical. If TRUE, only the commands will be written to
the commands_file. If FALSE, the commands will be executed.}

\item{metadata_table}{tsv table. if not specified, the file must be named
<EXPX>_metadata.tsv. This table must have: a column named "mjolnir_agnomens"
with the names given to the samples during the pipeline in FREYJA; a column
named "original_samples" with the samples names that will be given to the
samples at the end of the pipeline; and a column with the name specified in
in the "blank_col" parameter ("BLANK" by default) where blanks, negatives and
controls are tagged with a flag specified in the "blank_tag" parameter (T by
default).}
}
\description{
This function will prepare the FASTQ raw data for each sample with the sample
tags and primers removed.
The R1 output files will contain all forward sequences and
R2 the reverse sequences. Please, read the Details section before running.
}
\details{
This function considers the following scenarios:

- \bold{Scenario 1}. The samples are multiplexed in a library/libraries
   (it is possible to run multiple libraries
   at once from a single experiment).
   Each library has only two raw .fastq(.gz), R1 and R2.
   For each library there must be only one ngsfile (see below)
   and for the whole experiment only one metadata file (see below).
   mjolnir_agnomens (the standard name of the sample during the pipeline)
   must be unique, it must not be the same in different ngsfiles.
   \emph{lib_prefix} and \emph{R1_filenames} must be the same length.

- \bold{Scenario 2}. The samples are multiplexed but for
   each library we have more
   than two raw data files. For example, when your library has been
   sequenced across multiple lanes on a Novaseq. In this case, this function has
   to be run separately for each library and the option
   \emph{multilane} must be set to TRUE.
   In this case, \emph{lib_prefix} can only be
   of length one (only one library),
   but the \emph{R1_filenames} can be longer.

- \bold{Scenario 3}. The samples are demultiplexed but the primer remains.
   In this case this function will trim the primers and separate the sequences
   into fwd and rev files.


\bold{Starting with multiplexed libraries:}

Files required:

- \emph{ngsfilter file}. Needed only for multiplexed libraries
     For each library, a ngsfilter file is needed and must be named
     \bold{ngsfilter_<library identifier>.tsv}.
     This must contain five tab-separated
     columns and no header. The first column with the library identifier
     (four charachter identifier), the second with the mjolnir_agnomens,
     the third with the sample tags, the
     fourth with the forward primers and the fifth with the reverse primers.

- \emph{metadata file}.
     A metadata file containing at least two columns required,
     "original_samples" and "mjolnir_agnomens" ("fastq_name_R1" for
     demultiplexed libraries, see below), and named as
     \bold{<experiment identifier>_metadata.tsv}.

Important: when the same library has different sequencing lanes (NovaSeq),
this function has to be run separately for each library and the option
\emph{multilane} must be set to TRUE.


\bold{Starting with demultiplexed samples:}

If samples are already demultiplexed, primers need to be set
(\emph{primer_F} & \emph{primer_R}) or LERAY_XT primers
for COI will be used by default.

Files required:

- \emph{metadata file}. This function will read the
     names of each individual R1 fastq files (full name including extension)
     from a column in the \bold{<experiment identifier>_metadata.tsv}
     file, called "fastq_name_R1".
     In the metadata table, each sample in the "original_samples"
     column must have a their corresponding fastq_name_R1 and
     "mjolnir_agnomen" (LIBX_sample_XXX, i.e LIBA_sample_001).
     If the "fastq_name_R1" column is not in the metadata, This function will use the
     \emph{lib_prefix}.
}
\examples{
library(mjolnir)

# Define input fastq files (only names of R1 files are needed)
R1_filenames <- c("ULO1_R1.fastq.gz", "ULO2_R1.fastq.gz", "ULO3_R1.fastq.gz",
                  "ULO4_R1.fastq.gz")

# Input identifiers for the individual libraries to be used.
# It should be a 4-character name, matching the information in the
# ngsfilter files.
lib_prefix <- c("ULO1", "ULO2", "ULO3", "ULO4")

# experiment identifier
experiment <- 'ULOY'
# Enter number of cores to be used in parallel.
cores <- 7

mjolnir1_RAN(R1_filenames, lib_prefix = lib_prefix, experiment = experiment,
             cores = cores, R1_motif = "_R1", R2_motif = "_R2",
             tag_error = 0, primer_error = 0.1)
}
