% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjolnir2_FREYJA.R
\name{mjolnir2_FREYJA}
\alias{mjolnir2_FREYJA}
\title{FREYJA: Filtering of Reads, Enrollment, Yoke-reads Joining and Alignment}
\usage{
mjolnir2_FREYJA(
  experiment = NULL,
  cores = 1,
  Lmin = 299,
  Lmax = 320,
  min_overlap = 40,
  maxdiff = 0,
  error_rate = 0.5,
  R1_motif = "_R1",
  R2_motif = "_R2",
  commands_file = "commands_runned_FREYJA.txt",
  only_commands = FALSE,
  ...
)
}
\arguments{
\item{experiment}{Character string. Acronym for the experiment. This
acronym must be of 4 characters in capital letters. Do not mix up library and
experiment acronyms. However they can be the same.}

\item{cores}{Numeric. Number of threads for parallel processing.}

\item{Lmin}{Numeric. Minimum bp length for a sequence to be accepted.}

\item{Lmax}{Numeric. Maximum bp length for a sequence to be accepted.}

\item{min_overlap}{Numeric. Minimum overlap length between R1 and R2 reads.}

\item{maxdiff}{Numeric. Maximum number of differences allowed in the overlap.
If less than 1, it will be considered as a percentage of the overlap length.}

\item{error_rate}{Numeric. Maximum expected error rate allowed for the output
merged reads. For a given sequence, the expected error is the sum of error
probabilities for all the positions in the sequence. 
Since error probabilities can be small but not null, the expected error is
always greater than zero, and at most equal to the length of the sequence
when all positions in the sequence have an error probability of 1.0.}

\item{R1_motif}{Character string that distinguish the forward line file from
the reverse.}

\item{R2_motif}{Character string that distinguish the reverse line file from
the forward.}

\item{commands_file}{Character string. Name of the file where all commands
will be written. If NULL or missing, commands will not be recorded.}

\item{only_commands}{Logical. If TRUE, only the commands will be written to
the commands_file. If FALSE, the commands will be executed.}
}
\description{
FREYJA will use OBITools3 commands to merge paired-end reads, trim primer
sequences, filter by length, split sequences per sample and dereplicate
within each sample.
}
\details{
Input file fastq files are expected to be without primers sequence and all
forward sequences in the R1 file and all reverse sequences in the R2 file.
}
\examples{
library(mjolnir)

# Define input fastq files (only names of R1 files are needed)
R1_filenames <- c("ULO1_R1.fastq.gz", "ULO2_R1.fastq.gz", "ULO3_R1.fastq.gz",
                  "ULO4_R1.fastq.gz")

# Input identifiers for the individual libraries to be used. 
# It should be a 4-character name, matching the information in the 
# ngsfilter files.
lib_prefixes <- c("ULO1", "ULO2", "ULO3", "ULO4")

# experiment identifier
experiment <- 'ULOY'
# Enter number of cores to be used in parallel.
cores <- 7

mjolnir1_RAN(R1_filenames, lib_prefix = lib_prefixes, experiment = experiment,
             cores = cores, R1_motif = "_R1", R2_motif = "_R2")

# Run FREYJA
mjolnir2_FREYJA(experiment = experiment, cores = cores, Lmin=299, Lmax=320)
}
