% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjolnir7_LOKI.R
\name{mjolnir7_LOKI}
\alias{mjolnir7_LOKI}
\title{LOKI: LULU Overseeing with Kinship Identification}
\usage{
mjolnir7_LOKI(experiment = NULL, min_id = 0.84, discard = TRUE, ...)
}
\arguments{
\item{experiment}{Character string. Acronym for the experiment. This
acronym must be of 4 characters in capital letters. Do not mix up library and
experiment acronyms. However they can be the same.}

\item{min_id}{Numeric. Equivalent to the --id option from vsearch --usearch_global

From vsearch manual: Reject the sequence match if the pairwise identity is 
lower than min_id (value ranging from 0.0 to 1.0 included). The search process 
sorts target sequences by decreasing number of k-mers they hav e in common 
with the query sequence, using that information as a proxy for sequence 
similarity. That efficient pre-filtering also prevents pairwise alignments 
with weakly matching targets, as there needs to be at least 6 shared kmers 
to start the pairwise alignment, and at least one out of every 16 k-mers 
from the query needs to match the target. Consequently, using values lower 
than --id 0.5 is not likely to capture more weakly matching targets. The 
pairwise identity is by default defined as the number of 
(matching columns) / (alignment length - terminal gaps)}
}
\description{
LOKI is a convenient wrapper of LULU for the MJOLNIR3 metabarcoding pipeline.
}
\details{
LOKI starts from the combined dataset of abundances and taxonomy from the 
previous step (FRIGGA): <EXPX>_FRIGGA.tsv.
A match list of representative MOTU sequences is created using VSEARCH and saved as a txt file.
Then Units that are potential errors based on co-occurrence patterns
are labelled and removed using LULU.
The output is called <EXPX>_LOKI_Curated.tsv
}
\examples{
library(mjolnir)

# Define input fastq files (only names of R1 files are needed)
R1_filenames <- c("ULO1_R1.fastq.gz", "ULO2_R1.fastq.gz", "ULO3_R1.fastq.gz",
                  "ULO4_R1.fastq.gz")

# Input identifiers for the individual libraries to be used. 
# It should be a 4-character name, matching the information in the 
# ngsfilter files.
lib_prefixes <- c("ULO1", "ULO2", "ULO3", "ULO4")

# experiment identifier
experiment <- 'ULOY'
# Enter number of cores to be used in parallel.
cores <- 7

mjolnir1_RAN(R1_filenames, lib_prefix = lib_prefixes, experiment = experiment,
             cores = cores, R1_motif = "_R1", R2_motif = "_R2")

# Run FREYJA
mjolnir2_FREYJA(experiment = experiment, cores = cores, Lmin=299, Lmax=320)

# Run HELA
mjolnir3_HELA(experiment = experiment, cores = cores)

# Run ODIN
mjolnir4_ODIN(experiment = experiment, cores = cores, d = 13, 
              min_reads_MOTU = 2, min_reads_ESV = 2,
              min_relative = 1 / 50000, blank_relative = 0.1, 
              metadata_table = "", blank_col = "BLANK", blank_tag = TRUE, 
              alpha = 4, entropy = c(0.47, 0.23, 1.02, 313), 
              algorithm = "DnoisE_SWARM")

# set the directory where the database is stored
tax_db <- "~/taxo_NCBI/DUFA_COI"

# Run THOR
mjolnir5_THOR(experiment = experiment, cores = cores, 
              tax_db = tax_db, run_ecotag = T)

# Run FRIGGA
mjolnir6_FRIGGA(experiment = experiment)

# Run LOKI
mjolnir7_LOKI(experiment = experiment, min_id = .84)
}
