% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjolnir8_RAGNAROC.R
\name{numts}
\alias{numts}
\title{numts}
\usage{
numts(datas, is_metazoa = FALSE, motu, datas_length)
}
\arguments{
\item{is_metazoa}{Logical. whether if the MOTU is assigned to metazoa or not}

\item{motu}{Character. Name of the MOTU evaluated}

\item{datas_length}{Numeric vector. Vector with the number of characters in 
each sequence}

\item{data}{Data Frame of the sequences in the MOTU.}
}
\description{
Internal function of RAGNAROC to detect numt ESV within MOTUs
}
\details{
numt function will:
1: keep sequences of the same length as the seed
2: remove missaligned sequences
3: search for codon stops using all the mitochondrial genome codes with the 
Biostrings package and keep the one with less reads with stop codons. 
4: if the MOTU is assigned to metazoans and the seed has 313 bp then check it 
the following aminoacid positions have the conserved ones in metazoans:
20 = "H"; 23 ="G"; 32 = "N"; 81 ="D"; 95 ="G"
}
