% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjolnir3_HELA.R
\name{mjolnir3_HELA}
\alias{mjolnir3_HELA}
\title{HELA: Hierarchical Elimination of Lurking Artifacts}
\usage{
mjolnir3_HELA(lib, cores, vsearchpath = NULL)
}
\arguments{
\item{lib}{Character string. Acronym for the experiment. This
acronym must be of 4 characters in capital letters. Do not mix up library and
experiment acronyms. However they can be the same.}

\item{cores}{Numeric. Number of threads for parallel processing.}

\item{vsearchpath}{Character string specifying the PATH to the vsearch package.}
}
\description{
This function uses the uchime_denovo algorithm implemented in VSEARCH to 
remove chimaeric sequences from the dataset.
}
\details{
HELA works in a sample-by-sample basis. HELA will process all 
individual fasta files in the current folder 
matching the pattern EXPX_XXXX_sample_XXX.fasta being EXPX the acronym set 
by lib parameter. This allows for parallel 
computing, significantly decreasing calculation times.
}
\examples{
library(mjolnir)

# Define input fastq files (only names of R1 files are needed)
R1_filenames <-c("ULO1_R1.fastq.gz","ULO2_R1.fastq.gz","ULO3_R1.fastq.gz","ULO4_R1.fastq.gz")

# Input identifiers for the individual libraries to be used. It should be a 4-character name, matching the information in the ngsfilter files
lib_prefixes <- c("ULO1","ULO2","ULO3","ULO4")

# Enter number of cores to be used in parallel. 
cores <- 7

# Run RAN
mjolnir1_RAN(R1_filenames,cores,lib_prefixes,R1_motif="_R1",R2_motif="_R2")

# set experiment acronym
lib <- "ULOY"

# Run FREYJA
mjolnir2_FREYJA(lib_prefix = lib_prefixes,lib = lib,cores = cores,Lmin=299,Lmax=320)

# set the maximum number of cores possible
cores <- 16

# Run HELA
mjolnir3_HELA(lib,cores)
}
